pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadgrasp"
version = "0.1.0"
description = "Contact and Collision Aware Dexterous Grasping with Interaction Bisector Surface"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Zhiyuan Zhang", email = "zhiyuan@example.com"},
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["robotics", "grasping", "dexterous manipulation", "deep learning"]

dependencies = [
    # Core scientific computing
    "numpy>=1.23.0",
    "scipy>=1.10.0",
    
    # Deep learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "diffusers>=0.20.0",
    "einops>=0.6.0",
    
    # 3D processing
    "trimesh>=3.21.0",
    "open3d>=0.17.0",
    "transforms3d>=0.4.1",
    # pytorch3d requires special installation, see README
    "pytorch-kinematics>=0.5.0",
    
    # Robot model
    "ikpy>=3.3.0",
    "urdf_parser_py>=0.0.4",
    
    # Configuration
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    
    # Visualization
    "plotly>=5.14.0",
    "matplotlib>=3.7.0",
    "k3d>=2.15.0",
    
    # Utilities
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "termcolor>=2.0.0",
    "fire>=0.5.0",
    "ipdb>=0.13.0",
    "loguru>=0.7.0",
    
    # Logging
    "tensorboard>=2.12.0",
    "wandb>=0.15.0",
    
    # Image processing
    "opencv-python>=4.7.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.2.0",
    "pillow>=9.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
sim = [
    # IsaacGym must be installed manually from NVIDIA
]
sparse = [
    # MinkowskiEngine needs to be installed from source with CUDA
    # For CUDA 12.x compatibility, use the forked version:
    # cd thirdparty/MinkowskiEngine && pip install -e . --no-build-isolation
    # See: https://github.com/AzharSindhi/MinkowskiEngineCuda13
]
torchsdf = [
    # TorchSDF needs to be installed from source with CUDA:
    # cd thirdparty/TorchSDF && pip install -e . --no-build-isolation
    # See: https://github.com/wrc042/TorchSDF
]
pytorch3d = [
    # pytorch3d requires special installation:
    # pip install "git+https://github.com/facebookresearch/pytorch3d.git@stable"
    # or conda install pytorch3d -c pytorch3d
]
full = [
    "pyrender>=0.1.45",
    "pyglet>=2.0.0",
    "PyOpenGL>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/matthewmzy/CADGrasp"
Repository = "https://github.com/matthewmzy/CADGrasp"
Documentation = "https://github.com/matthewmzy/CADGrasp#readme"

[project.scripts]
cadgrasp-train = "cadgrasp.baseline.train:main"
cadgrasp-predict = "cadgrasp.evaluator.predict:main"
cadgrasp-ibs = "cadgrasp.ibs.compute:main"

[tool.setuptools]
package-dir = {"" = "src", "LASDiffusion" = "thirdparty/LASDiffusion"}

[tool.setuptools.packages.find]
where = ["src", "thirdparty"]
include = ["cadgrasp*", "LASDiffusion*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.urdf", "*.xml", "*.obj", "*.stl", "*.npy"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | thirdparty
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["thirdparty"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
